Fix error event so for firefox
This error only fires on firefox, and when it does it logs a console
error (because ‘event’ is not declared).

Also, note that load event is notoriously unreliable for images (Chrome
will not fire it if the image is loaded from cache).  Stay tuned next
for more reliable version of this code that will always fire the
complete.   For now this fixes the console error message in FF.